Implementation of Update Playlist WS OK

+++ WS/src/main/java/main/providers/youtube/YoutubeManager.java
@@ -1,13 +1,12 @@
 package main.providers.youtube;
 
 import java.util.ArrayList;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 import main.clients.IYoutubeClient;
+import main.clients.requestparams.VideoDetails;
 import main.clients.requestparams.VideoDuration;
 import main.exceptions.NotImplementedException;
 import main.model.youtube.playlist.PlaylistContentGroupInfo;
@@ -15,7 +14,9 @@ import main.model.youtube.playlist.PlaylistVideoType;
 import main.model.youtube.playlist.RetrievingSequenceType;
 import main.model.youtube.playlist.VideoItem;
 import main.providers.DataStorage;
+import main.utils.Converter;
 
+@Component
 public class YoutubeManager {
 	
 	@Autowired
@@ -30,7 +31,10 @@ public class YoutubeManager {
 	private final int DEFAULT_FORMULA = YoutubePlaylistComposer.FORMULA_BASIC;
 	private final String FORMULA_SELECTED_CONFIG_NAME = "youtubePlaylist_formulaSelected";
 	
-	private static final String CUSTOM_PLAYLIST_NAME = "CustomWakeUpPlaylist";
+	private static final String CUSTOM_PLAYLIST_NAME = "Custom WakeUp Playlist";
+	
+	private final int RETRIEVE_VIDEO_COUNT_LIMIT = 30;
+//	private final int RETRIEVE_VIDEO_COUNT_LIMIT = 7; // For tests purposes, I can use this line instead of the other one.
 	
 	
 	public YoutubeManager() {
@@ -52,8 +56,6 @@ public class YoutubeManager {
 		if(!YoutubePlaylistComposer.getFormulasCodesMapping().values().contains(formulaCode))
 			throw new Exception("No formula has a code that equals " + formulaCode);
 		
-		//TODO A continuer (utiliser un thread pour faire le boulot d'ajout à un moment donné) -> A réfléchir car si le traitement prends peu de temps, ça ne vaut surement pas le coup d'utiliser un thread et mieux vaut alors attendre de connaitre le résultat du traitement.
-		
 		// Obtenir la séquence de vidéos à intégrer dans la playlist
 		ArrayList<PlaylistContentGroupInfo> contentGroupInfo = this.playlistComposer.compose(formulaCode);
 		
@@ -65,8 +67,9 @@ public class YoutubeManager {
 
 		// Clean my custom playlist
 		String myCustomPlaylistId = this.youtubeClient.retrievePlaylistIDOfMyAccount(CUSTOM_PLAYLIST_NAME);
-		this.youtubeClient.DeletePlaylist(myCustomPlaylistId);
-		this.youtubeClient.CreatePlaylist(myCustomPlaylistId);
+		if(myCustomPlaylistId != null)
+			this.youtubeClient.DeletePlaylist(myCustomPlaylistId);
+		myCustomPlaylistId = this.youtubeClient.CreatePlaylist(CUSTOM_PLAYLIST_NAME);
 		
 		// Add videos to playlist
 		for (VideoItem video : videosToAdd) {
@@ -85,22 +88,28 @@ public class YoutubeManager {
 		PlaylistVideoType videoTypeInfo = videoSeqInfos.getTypeVideo();
 		RetrievingSequenceType videoSequenceTypeInfo = videoTypeInfo.getRetrievingSequenceInfo();
 		
-		VideoDuration videoDuration = VideoDuration.getVideoDurationForMaxMinutesValue(videoSeqInfos.getMaxMinutesForEachVideo());
+		VideoDuration videoDuration = videoSeqInfos.isTypeSearchAny() ? 
+				VideoDuration.Any 
+				: VideoDuration.getVideoDurationForMaxMinutesValue(videoSeqInfos.getMaxMinutesForEachVideo());
 		
 		ArrayList<VideoItem> result = null;
-		String channelId = null;
+		String channelId = videoTypeInfo.getChannelId() != null ? videoTypeInfo.getChannelId() : null;
 		switch(videoSequenceTypeInfo) {
 			case CHANNEL_LASTVIDEOS:
-				channelId = this.youtubeClient.retrieveChannelID(videoTypeInfo.getChannelName());
-				result = this.youtubeClient.listChannelVideos(channelId, videoDuration, Long.valueOf(7));
-				result = selectFilterVideos(result, videoSeqInfos, videoDuration);
+				if(channelId == null)
+					channelId = this.youtubeClient.retrieveChannelID(videoTypeInfo.getChannelName());
+				result = this.youtubeClient.listChannelVideos(channelId, videoDuration, Long.valueOf(RETRIEVE_VIDEO_COUNT_LIMIT));
+				fillMissingInformations(result);
+				result = selectFilterVideos(result, videoSeqInfos);
 				break;
 			case MUSIC_TRENDS: // Not possible to access the real music trends through the API, so we use a playlist management
 			case CHANNEL_PLAYLIST:
-				channelId = this.youtubeClient.retrieveChannelID(videoTypeInfo.getChannelName());
+				if(channelId == null)
+					channelId = this.youtubeClient.retrieveChannelID(videoTypeInfo.getChannelName());
 				String playlistId = this.youtubeClient.retrievePlaylistIDOfChannel(videoTypeInfo.getPlaylistName(), channelId);
-				result = this.youtubeClient.listChannelVideos(playlistId, videoDuration, Long.valueOf(7));
-				result = selectFilterVideos(result, videoSeqInfos, videoDuration);
+				result = this.youtubeClient.listPlaylistVideos(playlistId, Long.valueOf(RETRIEVE_VIDEO_COUNT_LIMIT));
+				fillMissingInformations(result);
+				result = selectFilterVideos(result, videoSeqInfos);
 				break;
 			case SIMPLE_RESEARCH:
 			default:
@@ -115,18 +124,38 @@ public class YoutubeManager {
 		return result;
 	}
 
+
+	private void fillMissingInformations(ArrayList<VideoItem> videoList) throws Exception {
+		for (VideoItem video : videoList) {
+			if(video.getDuration() == null || video.getCreationDate() == null) {
+				VideoDetails videoDetails = this.youtubeClient.retrieveDetailsOfVideo(video.getId());
+
+				if(videoDetails.videoDuration.isZero()) // A duration of Zero is a live video
+					continue;
+				
+				if(video.getDuration() == null)
+					video.setDuration(videoDetails.videoDuration);
+				if(video.getCreationDate() == null)
+					video.setCreationDate(videoDetails.creationDate);
+			}
+		}
+	}
+
 	// We don't have the duration information of the video in the research output of youtube API, so we use an alternative solution
-	private ArrayList<VideoItem> selectFilterVideos(ArrayList<VideoItem> videoList, PlaylistContentGroupInfo videoSeqInfos, VideoDuration videoDuration) {
+	private ArrayList<VideoItem> selectFilterVideos(ArrayList<VideoItem> videoList, PlaylistContentGroupInfo videoSeqInfos) {
 		int maxGroupDuration = videoSeqInfos.getMaxMinutesForGroup();
 		
 		ArrayList<VideoItem> result = new ArrayList<VideoItem>();
 		
-		int currentVideoDuration = 0;
+		double totalVideoDuration = 0;// In minutes
 		
 		for (VideoItem video : videoList) {
-			currentVideoDuration += videoDuration.getLessThanMinutes();
-			if(currentVideoDuration <= maxGroupDuration)
+			double currentVideoDuration = Converter.convertSecondsToMinutes(video.getDuration().getSeconds());
+			
+			if(totalVideoDuration + currentVideoDuration <= maxGroupDuration) {
 				result.add(video);
+				totalVideoDuration += Converter.convertSecondsToMinutes(video.getDuration().getSeconds());
+			}
 		}
 		
 		return result;
